test: 경로 관련 테스트 분리
